refactor(kirra-ota-ctl): split the CLI bin into per-subcommand modules
kirra_ota_ctl.rs was 1775 lines: eleven subcommands (run/stage/commit/rollback/
probe/pull/report/enroll/status/trust-provision/model-allowlist) plus their
option parsers, the shared HTTP/config plumbing, and two test modules, all in
one bin file. Convert to a src/bin/kirra_ota_ctl/ module dir, one concern per
submodule:

  main.rs   — Cfg + dispatch + the shared plumbing (http_get/http_download/
              http_post_json, now_ms, sign_report, exec_governor, set_executable,
              write_atomic, print_usage, load_release_pubkey, release verify)
  slots.rs  — run / stage / commit / rollback (A/B slot lifecycle)
  probe.rs  — health probe → commit-or-rollback (ProbeOpts)
  pull.rs   — campaign assignment pull + download (PullOpts)
  report.rs — adoption report (ReportOpts)
  enroll.rs — measured-boot enrollment (EnrollOpts) + its enroll_tests
  status.rs — slot status
  uptane.rs — Uptane trust provisioning + model allow-list + its cli tests

Behaviour-preserving relocation. Shared plumbing stays in the bin ROOT
(main.rs), where crate-root-private items are visible to every descendant
submodule — so those helpers need no widening. Only the `cmd_*` dispatch
targets are `pub(crate)` (the root's main() calls up into the submodules), plus
`stage_verified` + `DEFAULT_TRUST_STORE` which pull.rs shares. main() imports
the cmd_* by name. Cargo.toml's explicit bin `path` now points at main.rs.

The two test modules move with their concern (enroll_tests → enroll.rs,
model_allowlist_cli_tests → uptane.rs). Clean rebuild: bin compiles, 11 tests
pass, clippy clean, fmt clean. De-monolith #5 (final medium tier).
